XRDDEV-685 lift some exceptions to independent classes
